Split user and admin sesion detail page
