+++ ui/src/Routes.tsx
@@ -2,7 +2,7 @@ import { Navigate, Outlet, Route, Routes, useLocation } from 'react-router-dom';
 import HalfYearAttendances from './Attendance';
 import { Layout } from './Layout';
 import MyPage from './MyPage';
-import SessionDetail from './SessionDetail';
+import { AdminSessionDetail, UserSessionDetail } from './SessionDetail';
 import { AdminSessionList, UserSessionList } from './SessionList';
 import SignIn from './SignIn';
 import UserList from './UserList';
@@ -14,7 +14,7 @@ const AppRoutes = () => (
     <Route element={<Layout />}>
       <Route index element={<UserSessionList />} />
       <Route path="/sessions" element={<UserSessionList />} />
-      <Route path="/sessions/:id" element={<SessionDetail />} />
+      <Route path="/sessions/:id" element={<UserSessionDetail />} />
       <Route path="/users" element={<UserList />} />
       <Route element={<AuthRoute />}>
         <Route path="/me" element={<MyPage />} />
@@ -23,6 +23,7 @@ const AppRoutes = () => (
       <Route path="/signin" element={<SignIn />} />
       <Route path="/admin" element={<AdminRoute />}>
         <Route path="sessions" element={<AdminSessionList />} />
+        <Route path="sessions/:id" element={<AdminSessionDetail />} />
       </Route>
     </Route>
   </Routes>
