+++ ui/src/SessionDetail/admin/index.ts
@@ -1,0 +1,3 @@
+import AdminSessionDetail from './AdminSessionDetail';
+
+export default AdminSessionDetail;
