+++ ui/src/SessionDetail/user/index.ts
@@ -1,0 +1,3 @@
+import UserSessionDetail from './UserSessionDetail';
+
+export default UserSessionDetail;
